recognize all autoconf-standard --whateverdir options
plus program_prefix, program_suffix.

as in autoconf, most of these directories are defined,
by default, in terms of $prefix, $exec_prefix, and their
subdirectories.

as opposed to autoconf, the default values are expanded
eagerly, eg. when prefix='/snafu' then
includedir='/snafu/include' rather than '${prefix}/include'
as in autoconf.  the whole thing is a mother abyss of
turing tarpits, for the benefit of (quoting autoconf manual):

  This behavior is mandated by the GNU Coding Standards,
  so that when the user runs:

  ‘make’
      she can still specify a different prefix from the one
      specified to configure, in which case, if needed, the
      package should hard code dependencies corresponding
      to the make-specified prefix.

  ‘make install’
      she can specify a different installation location,
      in which case the package must still depend on the
      location which was compiled in (i.e., never recompile
      when ‘make install’ is run). This is an extremely
      important feature, as many people may decide to install
      all the files of a package grouped together, and then
      install links from the final locations to there.

the first benefit reads 'subvert build consistency' to me,
adds complexity most projects should not bear and, in practice
mostly do not actually support in their build scripts.

as for the second benefit, i'll be happy to leave the market for
this "extremely important feature" to autoconf.